projects controller tests

+++ server/src/api/projects/controller.ts
@@ -1,12 +1,13 @@
 import HTTP_STATUS from '../../constants/http-status'
 import { ApiTypes } from '../../types/api-types'
 import { ApiResponse } from '../../utils/api-response'
+import ApiError from '../../utils/error'
 import * as tasksService from '../tasks/service'
 import { ProjectCreate, ProjectCreatePublic, ProjectUpdate } from './schemas'
 import * as projectsService from './service'
 
 export async function addProject(req: ApiTypes.Request): Promise<ApiResponse> {
-  const userId = req.ctx!.decodedToken.user_id
+  const { user_id: userId } = req.ctx!.decodedToken
 
   const projectCreatePublic: ProjectCreatePublic = req.body
   const projectCreate: ProjectCreate = {
@@ -23,7 +24,7 @@ export async function addProject(req: ApiTypes.Request): Promise<ApiResponse> {
 export async function findByCreator(
   req: ApiTypes.Request
 ): Promise<ApiResponse> {
-  const userId = req.ctx!.decodedToken.user_id
+  const { user_id: userId } = req.ctx!.decodedToken
 
   const projects = await projectsService.findByCreator(userId)
   return new ApiResponse('Projects retreived', projects)
@@ -34,23 +35,41 @@ export async function deleteProject(
 ): Promise<ApiResponse> {
   const projectId: string = req.params.projectId
 
+  const { user_id: userId } = req.ctx!.decodedToken
+
+  const project = await projectsService.getProject(projectId)
+
+  if (project.creator != userId) {
+    throw new ApiError(
+      HTTP_STATUS.FORBIDDEN_403,
+      'You do not have permissions to delete this project'
+    )
+  }
+
   const projectPromise = projectsService.deleteProject(projectId)
   const tasksPromise = tasksService.deleteByProject(projectId)
 
-  const [project] = await Promise.all([projectPromise, tasksPromise])
+  const [projectDeleted] = await Promise.all([projectPromise, tasksPromise])
 
-  return new ApiResponse('Project deleted', project)
+  return new ApiResponse('Project deleted', projectDeleted)
 }
 
 export async function updateProject(
   req: ApiTypes.Request
 ): Promise<ApiResponse> {
-  const userId = req.ctx!.decodedToken.user_id
+  const { user_id: userId } = req.ctx!.decodedToken
 
   const projectId = req.params.projectId
   const projectUpdate: ProjectUpdate = req.body
 
   const project = await projectsService.getProject(projectId)
+
+  if (project.creator != userId) {
+    throw new ApiError(
+      HTTP_STATUS.FORBIDDEN_403,
+      'You do not have permissions to update this project'
+    )
+  }
 
   if (projectUpdate.name && projectUpdate.name !== project.name) {
     await projectsService.projectNameExists(userId, projectUpdate.name)
